services/ml-training/pyproject.toml
[tool.poetry]
name = "ml-training"
version = "0.1.0"
description = ""
authors = ["ray <wuhaohsiang1992@gmail.com>"]
readme = "README.md"
packages = [{include = "ml_training"}]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
scikit-learn = "1.1.1"
minio = "^7.1.13"
fastapi = "^0.93.0"
psycopg2-binary = "^2.9.5"
pandas = "^1.5.3"
rich = "^13.3.2"
uvicorn = "^0.20.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
